cc_search options

cc_search options

cc_params

Format

An object of class list of length 9.

Details

cc_params is a list with slots matching the parameters in cc_search - so you can more easily get to the options you want to pass in to that function.

The options are also listed below.

file set options

12: Delta Method IPCC AR5

4: Delta Method IPCC AR4

9: MarkSim Pattern Scaling

10: Eta South America

7: PRECIS Andes

8: CORDEX

11: Disaggregation IPCC AR4

3: Delta Climgen

2: Delta Method IPCC AR4 (Climgen Data)

5: Delta Method IPCC AR4 (Stanford Data)

6: Delta Method IPCC AR3

scenario options

1: Baseline

2: SRES A1B

3: SRES A2A

4: SRES B2A

5: SRES A2

6: SRES B1

7: RCP 2.6

8: RCP 4.5

9: RCP 6.0

10: RCP 8.5

model options

1: baseline

42: bcc_csm1_1

43: bcc_csm1_1_m

2: bccr_bcm2_0

44: bnu_esm

45: cccma_cancm4

46: cccma_canesm2

3: cccma_cgcm2_0

4: cccma_cgcm3_1_t47

5: cccma_cgcm3_1_t63

47: cesm1_bgc

48: cesm1_cam5

49: cesm1_cam5_1_fv2

50: cesm1_fastchem

51: cesm1_waccm

52: cmcc_cesm

53: cmcc_cm

54: cmcc_cms

6: cnrm_cm3

55: cnrm_cm5

56: csiro_access1_0

57: csiro_access1_3

7: csiro_mk2

8: csiro_mk3_0

25: csiro_mk3_5

58: csiro_mk3_6_0

59: ec_earth

60: fio_esm

9: gfdl_cm2_0

61: gfdl_cm2_1

10: gfdl_cm2_1

62: gfdl_cm3

63: gfdl_esm2g

64: gfdl_esm2m

11: giss_aom

65: giss_e2_h

66: giss_e2_h_cc

67: giss_e2_r

68: giss_e2_r_cc

26: giss_model_eh

27: giss_model_er

38: hadcm_cntrl

39: hadcm_high

40: hadcm_low

41: hadcm_midi

12: hccpr_hadcm3

13: iap_fgoals1_0_g

28: ingv_echam4

37: inm_cm3_0

69: inm_cm4

14: ipsl_cm4

70: ipsl_cm5a_lr

71: ipsl_cm5a_mr

72: ipsl_cm5b_lr

73: lasg_fgoals_g2

74: lasg_fgoals_s2

15: miroc3_2_hires

16: miroc3_2_medres

75: miroc_esm

76: miroc_esm_chem

77: miroc_miroc4h

78: miroc_miroc5

17: miub_echo_g

34: mohc_hadam3p_2

35: mohc_hadam3p_3

79: mohc_hadcm3

31: mohc_hadcm3q0

33: mohc_hadcm3q16

32: mohc_hadcm3q3

80: mohc_hadgem2_cc

81: mohc_hadgem2_es

29: mpi_echam4

18: mpi_echam5

30: mpi_echam5

82: mpi_esm_lr

83: mpi_esm_mr

84: mpi_esm_p

19: mri_cgcm2_3_2a

85: mri_cgcm3

20: ncar_ccsm3_0

86: ncar_ccsm4

21: ncar_pcm1

87: ncc_noresm1_m

88: ncc_noresm1_me

36: ncep_r2

22: nies99

89: nimr_hadgem2_ao

23: ukmo_hadcm3

24: ukmo_hadgem1

period options

1: 1970s

10: 1990s

2: 2000s

3: 2020s

4: 2030s

5: 2040s

6: 2050s

7: 2060s

8: 2070s

9: 2080s

variable options

1: Bioclimatics

6: Diurnal Temperature Range

3: Maximum Temperature

4: Mean Temperature

5: Minimum Temperature

2: Precipitation

7: Solar Radiation

9999: Other

resolution options

1: 30 seconds

2: 2.5

3: 5 minutes

4: 10 minutes

5: 30 minutes

6: 25 minutes

7: 20 minutes